feat: Remove backward compatibility and replace legacy auth functions
- Remove legacy auth.go and auth_test.go files
- Replace InternalAuthMiddleware with AuthMiddleware
- Replace InternalRequirePermission with RequirePermission
- Replace GetInternalUserFromContext with GetUserFromContext
- Replace InternalUserOwnsSession with UserOwnsSession
- Update notification service to directly use internal entities.User
- Remove backward compatibility conversion functions
- Rename internal_middleware.go to middleware.go
- Remove unused contextKey and constants from github.go
- Clean up all legacy test files that used old APIs

All auth functions now directly use internal entities without compatibility layer.
